refactor(type): split combat damage result types
