+++ Source/Portfolio/Type/CCombatHitTypes.cpp
@@ -1,0 +1,6 @@
+#include "Type/CCombatHitTypes.h"
+
+bool FOverlapContext::IsValidMinimal() const
+{
+	return IsValid(OwnerActor) && IsValid(DamageCauser) && IsValid(OtherActor);
+}
