+++ Source/Portfolio/Type/CCombatResultTypes.cpp
@@ -1,0 +1,1 @@
+#include "Type/CCombatResultTypes.h"
